chore: WIP cleanup webauthn
