+++ packages/core/src/lib/actions/callback/index.ts
@@ -23,7 +23,11 @@ import type {
   User,
 } from "../../../types.js"
 import type { Cookie, SessionStore } from "../../utils/cookie.js"
-import { assertInternalOptionsWebAuthn, verifyAuthenticate, verifyRegister } from "../../utils/webauthn-utils.js"
+import {
+  assertInternalOptionsWebAuthn,
+  verifyAuthenticate,
+  verifyRegister,
+} from "../../../providers/webauthn.js"
 
 /** Handle callbacks from login services */
 export async function callback(
@@ -370,7 +374,10 @@ export async function callback(
     } else if (provider.type === "webauthn" && method === "POST") {
       // Get callback action from request. It should be either "authenticate" or "register"
       const action = request.body?.action
-      if (typeof action !== "string" || (action !== "authenticate" && action !== "register")) {
+      if (
+        typeof action !== "string" ||
+        (action !== "authenticate" && action !== "register")
+      ) {
         throw new AuthError("Invalid action parameter")
       }
       // Return an error if the adapter is missing or if the provider
@@ -383,7 +390,11 @@ export async function callback(
       let authenticator: Authenticator | undefined
       switch (action) {
         case "authenticate": {
-          const verified = await verifyAuthenticate(localOptions, request, cookies)
+          const verified = await verifyAuthenticate(
+            localOptions,
+            request,
+            cookies
+          )
 
           user = verified.user
           account = verified.account
@@ -402,13 +413,15 @@ export async function callback(
       }
 
       // Check if user is allowed to sign in
-      await handleAuthorized(
-        { user, account },
-        options,
-      )
+      await handleAuthorized({ user, account }, options)
 
       // Sign user in, creating them and their account if needed
-      const { user: loggedInUser, isNewUser, session, account: currentAccount } = await handleLoginOrRegister(
+      const {
+        user: loggedInUser,
+        isNewUser,
+        session,
+        account: currentAccount,
+      } = await handleLoginOrRegister(
         sessionStore.value,
         user,
         account,
@@ -422,7 +435,10 @@ export async function callback(
 
       // Create new authenticator if needed
       if (authenticator && loggedInUser.id) {
-        await localOptions.adapter.createAuthenticator({ ...authenticator, userId: loggedInUser.id })
+        await localOptions.adapter.createAuthenticator({
+          ...authenticator,
+          userId: loggedInUser.id,
+        })
       }
 
       // Do the session registering dance
@@ -470,15 +486,20 @@ export async function callback(
         })
       }
 
-      await events.signIn?.({ user: loggedInUser, account: currentAccount, isNewUser })
+      await events.signIn?.({
+        user: loggedInUser,
+        account: currentAccount,
+        isNewUser,
+      })
 
       // Handle first logins on new accounts
       // e.g. option to send users to a new account landing page on initial login
       // Note that the callback URL is preserved, so the journey can still be resumed
       if (isNewUser && pages.newUser) {
         return {
-          redirect: `${pages.newUser}${pages.newUser.includes("?") ? "&" : "?"
-            }${new URLSearchParams({ callbackUrl })}`,
+          redirect: `${pages.newUser}${
+            pages.newUser.includes("?") ? "&" : "?"
+          }${new URLSearchParams({ callbackUrl })}`,
           cookies,
         }
       }
