+++ packages/core/src/lib/actions/callback/oauth/checks.ts
@@ -13,7 +13,7 @@ import type { Cookie } from "../../../utils/cookie.js"
 import type { OAuthConfigInternal } from "../../../../providers/oauth.js"
 import type { WebAuthnProviderType } from "../../../../providers/webauthn.js"
 
-interface CheckPayload {
+export interface CheckPayload {
   value: string
 }
 
@@ -263,53 +263,4 @@ export function handleState(
   }
 
   return { randomState, proxyRedirect }
-}
-
-const WEBAUTHN_CHALLENGE_MAX_AGE = 60 * 15 // 15 minutes in seconds
-type WebAuthnChallengeCookie = { challenge: string; registerData?: User }
-export const webauthnChallenge = {
-  async create(options: InternalOptions<WebAuthnProviderType>, challenge: string, registerData?: User) {
-    const maxAge = WEBAUTHN_CHALLENGE_MAX_AGE
-    const data: WebAuthnChallengeCookie = { challenge, registerData }
-    const cookie = await signCookie(
-      "webauthnChallenge",
-      JSON.stringify(data),
-      maxAge,
-      options
-    )
-    return { cookie }
-  },
-
-  /**
-   * Returns challenge if present,
-   */
-  async use(
-    options: InternalOptions<WebAuthnProviderType>,
-    cookies: RequestInternal["cookies"],
-    resCookies: Cookie[],
-  ): Promise<WebAuthnChallengeCookie> {
-    const challenge = cookies?.[options.cookies.webauthnChallenge.name]
-
-    if (!challenge)
-      throw new InvalidCheck("Challenge cookie missing.")
-
-    const value = await decode<CheckPayload>({
-      ...options.jwt,
-      token: challenge,
-      salt: options.cookies.webauthnChallenge.name,
-    })
-
-    if (!value?.value)
-      throw new InvalidCheck("Challenge value could not be parsed.")
-
-    // Clear the pkce code verifier cookie after use
-    const cookie = {
-      name: options.cookies.webauthnChallenge.name,
-      value: "",
-      options: { ...options.cookies.webauthnChallenge.options, maxAge: 0 },
-    }
-    resCookies.push(cookie)
-
-    return JSON.parse(value.value) as WebAuthnChallengeCookie
-  },
 }
