+++ packages/core/src/lib/utils/webauthn-client.js
@@ -1,224 +1,165 @@
-//@ts-check
+// This is the JS version of webauthn-client.ts. It is inlined in the HTML file so it needs to be JS.
+// TODO: Generate programatically from webauthn-client.ts
 
-// Declare a SimpleWebAuthnBrowser variable as part of "window"
+// @ts-check
 
-/** @typedef {"authenticate"} WebAuthnAuthenticate */
-/** @typedef {"register"} WebAuthnRegister */
-/** @typedef {WebAuthnRegister | WebAuthnAuthenticate} WebAuthnOptionsAction */
 /**
- * @template {WebAuthnOptionsAction} T
- * @typedef {T extends WebAuthnAuthenticate ?
- *  { options: import("@simplewebauthn/server/script/deps").PublicKeyCredentialRequestOptionsJSON; action: "authenticate" } :
- *  T extends WebAuthnRegister ?
- *  { options: import("@simplewebauthn/server/script/deps").PublicKeyCredentialCreationOptionsJSON; action: "register" } :
- * never
- * } WebAuthnOptionsReturn
+ * @typedef {import("../../providers/webauthn.js").WebAuthnAction} WebAuthnAction
+ * @typedef {import("../../providers/webauthn.js").WebAuthnOptionsResponseBody} WebAuthnOptionsResponseBody
+ * @typedef {import("../../providers/webauthn.js").WebAuthnRegister} WebAuthnRegister
+ * @typedef {import("../../providers/webauthn.js").WebAuthnAuthenticate} WebAuthnAuthenticate
  */
 
 /**
- * webauthnScript is the client-side script that handles the webauthn form
+ * The client-side script that handles the WebAuthn form.
  * 
- * @param {string} authURL is the URL of the auth API
- * @param {string} providerID is the ID of the webauthn provider
+ * @param {string} authURL - The URL of the auth API.
+ * @param {string} providerID - The ID of the webauthn provider.
  */
 export async function webauthnScript(authURL, providerID) {
-  /** @type {typeof import("@simplewebauthn/browser")} */
-  // @ts-ignore
-  const WebAuthnBrowser = window.SimpleWebAuthnBrowser
+  // @ts-expect-error
+  const WebAuthnBrowser = window.SimpleWebAuthnBrowser;
 
   /**
-   * Fetch webauthn options from the server
+   * Fetch WebAuthn options from the server.
    * 
-   * @template {WebAuthnOptionsAction} T
-   * @param {T | undefined} action action to fetch options for
-   * @returns {Promise<WebAuthnOptionsReturn<T> | undefined>}
+   * @template {string | undefined} T
+   * @param {T} action - Action to fetch options for.
+   * @returns {Promise<T extends "authenticate" ? WebAuthnAuthenticate : T extends "register" ? WebAuthnRegister : WebAuthnAuthenticate | WebAuthnRegister>}
    */
   async function fetchOptions(action) {
-    // Create the options URL with the action and query parameters
-    const url = new URL(`${authURL}/webauthn-options/${providerID}`)
+    const url = new URL(`${authURL}/webauthn-options/${providerID}`);
 
-    if (action) url.searchParams.append("action", action)
+    if (action) url.searchParams.append("action", action);
 
-    const formFields = getFormFields()
-    formFields.forEach((field) => {
+    const formFields = getFormFields();
+    formFields.forEach((field) =>
       url.searchParams.append(field.name, field.value)
-    })
+    );
 
-    const res = await fetch(url)
-    if (!res.ok) {
-      console.error("Failed to fetch options", res)
-
-      return
-    }
-
-    return res.json()
+    const res = await fetch(url);
+    if (res.ok) return res.json();
+    throw new Error("Failed to fetch options");
   }
 
   /**
-   * Get the webauthn form from the page
+   * Get the WebAuthn form from the page.
    * 
    * @returns {HTMLFormElement}
    */
   function getForm() {
-    const formID = `#${providerID}-form`
+    const formID = `#${providerID}-form`;
     /** @type {HTMLFormElement | null} */
-    const form = document.querySelector(formID)
-    if (!form) throw new Error(`Form '${formID}' not found`)
-
-    return form
+    const form = document.querySelector(formID);
+    if (form) return form;
+    throw new Error(`Form '${formID}' not found`);
   }
 
   /**
-   * Get formFields from the form
+   * Get formFields from the form.
    * 
    * @returns {HTMLInputElement[]}
    */
   function getFormFields() {
-    const form = getForm()
+    const form = getForm();
     /** @type {HTMLInputElement[]} */
-    const formFields = Array.from(form.querySelectorAll("input[data-form-field]"))
+    const formFields = Array.from(form.querySelectorAll("input[data-form-field]"));
 
-    return formFields
+    return formFields;
   }
 
   /**
    * Passkey form submission handler.
-   * Takes the input from the form and a few other parameters and submits it to the server.
    * 
-   * @param {WebAuthnOptionsAction} action action to submit
-   * @param {unknown | undefined} data optional data to submit
-   * @returns {Promise<void>}
+   * @param {WebAuthnAction} action - Action to submit.
+   * @param {unknown} data - Optional data to submit.
    */
   async function submitForm(action, data) {
-    const form = getForm()
+    const form = getForm();
 
-    // If a POST request, create hidden fields in the form
-    // and submit it so the browser redirects on login
     if (action) {
-      const actionInput = document.createElement("input")
-      actionInput.type = "hidden"
-      actionInput.name = "action"
-      actionInput.value = action
-      form.appendChild(actionInput)
+      const actionInput = document.createElement("input");
+      actionInput.type = "hidden";
+      actionInput.name = "action";
+      actionInput.value = action;
+      form.appendChild(actionInput);
     }
 
     if (data) {
-      const dataInput = document.createElement("input")
-      dataInput.type = "hidden"
-      dataInput.name = "data"
-      dataInput.value = JSON.stringify(data)
-      form.appendChild(dataInput)
+      const dataInput = document.createElement("input");
+      dataInput.type = "hidden";
+      dataInput.name = "data";
+      dataInput.value = JSON.stringify(data);
+      form.appendChild(dataInput);
     }
 
-    return form.submit()
+    return form.submit();
   }
 
   /**
-   * Executes the authentication flow by fetching options from the server,
-   * starting the authentication, and submitting the response to the server.
+   * Executes the authentication flow.
    * 
-   * @param {WebAuthnOptionsReturn<WebAuthnAuthenticate>['options']} options
-   * @param {boolean} autofill Whether or not to use the browser's autofill
-   * @returns {Promise<void>}
+   * @param {WebAuthnAuthenticate} options
+   * @param {boolean} autofill - Whether or not to use the browser's autofill.
    */
   async function authenticationFlow(options, autofill) {
-    // Start authentication
-    const authResp = await WebAuthnBrowser.startAuthentication(options, autofill)
-
-    // Submit authentication response to server
-    return await submitForm("authenticate", authResp)
+    const authResp = await WebAuthnBrowser.startAuthentication(options, autofill);
+    return await submitForm("authenticate", authResp);
   }
 
   /**
-   * @param {WebAuthnOptionsReturn<WebAuthnRegister>['options']} options
+   * Executes the registration flow.
+   * 
+   * @param {WebAuthnRegister} options
    */
   async function registrationFlow(options) {
-    // Check if all required formFields are set
-    const formFields = getFormFields()
+    const formFields = getFormFields();
     formFields.forEach((field) => {
       if (field.required && !field.value) {
-        throw new Error(`Missing required field: ${field.name}`)
+        throw new Error(`Missing required field: ${field.name}`);
       }
-    })
+    });
 
-    // Start registration
-    const regResp = await WebAuthnBrowser.startRegistration(options)
-
-    // Submit registration response to server
-    return await submitForm("register", regResp)
+    const regResp = await WebAuthnBrowser.startRegistration(options);
+    return await submitForm("register", regResp);
   }
 
   /**
-   * Attempts to authenticate the user when the page loads
-   * using the browser's autofill popup.
-   * 
-   * @returns {Promise<void>}
+   * Attempts to authenticate the user using the browser's autofill popup on page load.
    */
   async function autofillAuthentication() {
-    // if the browser can't handle autofill, don't try
-    if (!WebAuthnBrowser.browserSupportsWebAuthnAutofill()) return
-
-    const res = await fetchOptions("authenticate")
-    if (!res) {
-      console.error("Failed to fetch option for autofill authentication")
-
-      return
-    }
-
+    if (!WebAuthnBrowser.browserSupportsWebAuthnAutofill()) return;
     try {
-      await authenticationFlow(res.options, true)
+      const res = await fetchOptions("authenticate");
+      await authenticationFlow(res, true);
     } catch (e) {
-      console.error(e)
+      console.error(e);
     }
   }
 
   /**
-   * Sets up the passkey form by overriding the form submission handler
-   * so that it attempts to authenticate the user when the form is submitted.
-   * If the user is not registered, it will attempt to register them instead.
+   * Sets up the passkey form by overriding the form submission handler.
    */
   async function setupForm() {
-    const form = getForm()
+    const form = getForm();
 
-    // If the browser can't do WebAuthn, hide the form
     if (!WebAuthnBrowser.browserSupportsWebAuthn()) {
-      form.style.display = "none"
-
-      return
+      form.style.display = "none";
+      return;
     }
 
-    if (form) {
-      form.addEventListener("submit", async (e) => {
-        e.preventDefault()
-
-        // Fetch options from the server without assuming that
-        // the user is registered
-        const res = await fetchOptions(undefined)
-        if (!res) {
-          console.error("Failed to fetch options for form submission")
-
-          return
-        }
-
-        // Then execute the appropriate flow
-        if (res.action === "authenticate") {
-          try {
-            await authenticationFlow(res.options, false)
-          } catch (e) {
-            console.error(e)
-          }
-        } else if (res.action === "register") {
-          try {
-            await registrationFlow(res.options)
-          } catch (e) {
-            console.error(e)
-          }
-        }
-      })
-    }
+    form.addEventListener("submit", async (e) => {
+      e.preventDefault();
+      try {
+        const res = await fetchOptions(undefined);
+        if (res.action === "authenticate") await authenticationFlow(res, false);
+        else await registrationFlow(res);
+      } catch (e) {
+        console.error(e);
+      }
+    });
   }
 
-  // On page load, setup the form and attempt to authenticate the user.
-  setupForm()
-  autofillAuthentication()
+  setupForm();
+  autofillAuthentication();
 }
