+++ packages/next-auth/src/webauthn.ts
@@ -3,7 +3,7 @@ import { startAuthentication, startRegistration } from "@simplewebauthn/browser"
 import { getCsrfToken, getProviders, __NEXTAUTH } from "./react.js"
 
 import type { LoggerInstance } from "@auth/core/types"
-import type { WebAuthnOptionsResponseBody } from "@auth/core/lib/utils/webauthn-utils"
+import type { WebAuthnOptionsResponseBody } from "@auth/core/types"
 import type {
   BuiltInProviderType,
   RedirectableProviderType,
@@ -49,10 +49,10 @@ async function webAuthnOptions(
   const optionsData: WebAuthnOptionsResponseBody = await optionsResp.json()
 
   if (optionsData.action === "authenticate") {
-    const webAuthnResponse = await startAuthentication(optionsData.options)
+    const webAuthnResponse = await startAuthentication(optionsData)
     return { data: webAuthnResponse, action: "authenticate" }
   } else {
-    const webAuthnResponse = await startRegistration(optionsData.options)
+    const webAuthnResponse = await startRegistration(optionsData)
     return { data: webAuthnResponse, action: "register" }
   }
 }
